[R] Moved Currency and HTTP helper classes to other packages
Currency.java from com.example.currencyconverter.util.enums
to com.example.currencyconverter.models

HttpHelper.java from com.example.currencyconverter.util.web
to com.example.currencyconverter.web

Removed the util package and its nested packages
